requirements.txt
paho-mqtt>=2.0.0
anthropic>=0.40.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
pygame>=2.5.0
requests>=2.31.0
